added modal to userInfo in dashboard
