+++ src/views/ReserveOverview/UserInfoRow.tsx
@@ -7,6 +7,7 @@ interface Props {
   value: any;
   type?: string;
   enableModal?: boolean;
+  modalOpen?: any
 }
 
 const UserInfoRow: React.FC<Props> = ({ ...props }) => {
@@ -26,7 +27,7 @@ const UserInfoRow: React.FC<Props> = ({ ...props }) => {
         <Flex>
           <Text fontSize="2xl">{title}</Text>
           {props.enableModal && (
-            <ModalIcon position="relative" top="0" left="2" maxHeight="1.2rem" onOpen={() => {}} />
+            <ModalIcon position="relative" top="0" left="2" maxHeight="1.2rem" onOpen={() => props.modalOpen(props.title)} />
           )}
         </Flex>
         <Flex>
